requirements.txt
biosimulators_utils >= 0.1.32
numpy
python_copasi >= 4.30.233
